feat: support llama completion provider options
